requirements.txt
imgurpython==1.1.7
praw==7.1.0
Pillow==4.1.1
